Add component CardEmotion + function to fetch emotion (not working without server)
