+++ interface-web/src/Constants/constantsEmotions.tsx
